Add files via upload
Project 1- Python College Database 
It is a command-line project where the developer’s goal is to build a student database application. 
The users can make the use of such Python applications to save the contact details of the student. 
The users are also allowed to find the contact details, update or delete saved contacts details and display them using GUI.
 
Project 2- Scientific GUI Calculator using Tkinter in Python with trigonometry functions.